refactor: move initial place load in main page, refactor components to use new places store
