+++ src/shared/stores/places/index.ts
@@ -1,2 +1,3 @@
 export * from './hooks';
 export * from './actions';
+export * from './types';
