+++ src/widgets/Map/types/index.ts
@@ -1,0 +1,5 @@
+import { type PlacesDataWithGeo } from 'features/LoadMap/types';
+
+export interface MainMapProps {
+  placesGeo: PlacesDataWithGeo;
+}
