+++ src/widgets/PlacesList/types/index.ts
@@ -1,0 +1,5 @@
+import { type Place } from 'shared/stores/places';
+
+export interface PlacesListProps {
+  places: Place[];
+}
