+++ src/widgets/PlacesList/ui/PlacesList.tsx
@@ -1,26 +1,23 @@
 import { useLocation } from 'react-router-dom';
+import { type PlacesListProps } from 'widgets/PlacesList/types';
 import { PlaceCard } from 'features/PlaceCard';
-import { usePlaces } from 'shared/context/PlacesData/usePlaces';
-import { usePlacesStore } from 'shared/stores/places';
+import { setShowFavorites, usePlacesStore } from 'shared/stores/places';
 import cls from './PlacesList.module.scss';
 
-export function PlacesList() {
+export function PlacesList({ places }: PlacesListProps) {
   const location = useLocation();
-  const { filterablePlaces, favoritePlaces, showFavorites } = usePlaces();
-  const { setShowFavorite } = usePlaces();
 
-  const isFiltered = usePlacesStore((state) => state.isFiltered);
-
-  if (!filterablePlaces.length) return null;
+  const showFavorites = usePlacesStore((state) => state.showFavorites);
+  const filteredPlaces = usePlacesStore((state) => state.filteredPlaces);
 
   return (
-    !isFiltered && (
+    !filteredPlaces && (
       <>
         <div
-          className={`${cls.placesData} ${showFavorites && favoritePlaces?.length && location.pathname !== '/details' ? cls.showFavorites : ''}`}
+          className={`${cls.placesData} ${showFavorites && location.pathname !== '/details' ? cls.showFavorites : ''}`}
         >
           <div className={`${cls.PlacesList} ${location.pathname === '/details' ? cls.detailsOpen : ''}`}>
-            {(showFavorites && favoritePlaces?.length ? favoritePlaces : filterablePlaces).map((place) => (
+            {places.map((place) => (
               <PlaceCard
                 properties={place.properties}
                 coordinates={place.geometry.coordinates}
@@ -32,7 +29,7 @@ export function PlacesList() {
         <div className={cls.backdrop}>
           <div
             onClick={() => {
-              setShowFavorite(false);
+              setShowFavorites(false);
             }}
             className={cls.closeButton}
           ></div>
